FIx test compilation
