+++ enclone_exec/Cargo.toml
@@ -27,6 +27,7 @@ path = "src/bin/enclone.rs"
 chrono = { version = "0.4", default-features = false, features = ["std", "clock"] }
 enclone_core = { git = "https://github.com/10XGenomics/enclone_ranger.git", branch = "main" }
 enclone_main = { path = "../enclone_main" }
+enclone_testlist = { path = "../enclone_testlist" }
 enclone_visual = { path = "../enclone_visual", optional=true }
 flate2 = "1"
 home = "0.5"
