+++ enclone_paper/Cargo.toml
@@ -26,6 +26,7 @@ bio_edit = { version = "0.1", git = "https://github.com/10XGenomics/rust-toolbox
 debruijn = "0.3"
 enclone_core = { git = "https://github.com/10XGenomics/enclone_ranger.git", branch = "main" }
 enclone_tail = { path = "../enclone_tail" }
+enclone_testlist = { path = "../enclone_testlist" }
 flate2 = "1"
 io_utils = { version = "0.3", git = "https://github.com/10XGenomics/rust-toolbox.git", branch = "master" }
 itertools = "0.10"
