+++ enclone_tail/src/group.rs
@@ -22,6 +22,7 @@ use ansi_escape::ansi_to_html::{
 };
 use ansi_escape::{emit_bold_escape, emit_eight_bit_color_escape, emit_end_escape};
 use enclone_build::version_string;
+use enclone_core::barcode_fate::BarcodeFate;
 use enclone_core::combine_group_pics::combine_group_pics;
 use enclone_core::defs::{ColInfo, EncloneControl, ExactClonotype, GexInfo};
 use enclone_core::mammalian_fixed_len::mammalian_fixed_len_peer_groups;
@@ -56,7 +57,7 @@ pub fn group_and_print_clonotypes(
     join_info: &Vec<(usize, usize, bool, Vec<u8>)>,
     gex_info: &GexInfo,
     vdj_cells: &Vec<Vec<String>>,
-    fate: &Vec<HashMap<String, &'static str>>,
+    fate: &Vec<HashMap<String, BarcodeFate>>,
     dref: &Vec<DonorReferenceItem>,
     groups: &Vec<Vec<(i32, String)>>,
     opt_d_val: &Vec<(usize, Vec<Vec<Vec<usize>>>)>,
