+++ enclone_testlist/src/enclone.testdata
@@ -1,0 +1,128 @@
+# These are human BCR data that are in general not going to be released publicly, either because 
+# they are not unambiguously consented for public release, or because they contain data from
+# experimental conditions (typically titrations of reagents) and would be confusing.
+#
+# Each non-comment line in this file is a list of sample IDs, putatively from the same donor.
+
+40043 
+
+40086 
+
+83817,83819,83821
+
+# This died when upgrading to 3.1, with:
+# No input FASTQs were found with the requested sample indices.
+
+# 90514 
+
+91673 
+
+# This died when upgrading to 3.1, with:
+# /mnt/analysis/marsoc/pipestances/HJVNVBBXX/BCL_PROCESSOR_PD/HJVNVBBXX/1015.12.3-0/BCL_PROCESSOR_PD/BCL_PROCESSOR/DEMULTIPLEX/fork0/files/demultiplexed_fastq_path: no such file or directory.
+
+# 39860
+
+# ChronicLymphLeuk.
+
+48620,48634
+
+# Lupus.  These look like they're from two separate donors.
+
+48622 
+83812,83813,86227,86228,86237,86238
+
+# multiple sclerosis
+
+48618 
+
+48614 
+
+45987,45977
+
+# MALT
+
+83808-83809,86229-86230,86233-86234,89644-89645
+
+47680,47674
+
+# rheumatoid arthritis
+
+86225,86226,89642,89643
+
+# the following appears to share with multiple myeloma, will need to investigate
+
+48612,48630
+
+# multiple myeloma.  Maybe can combine.
+
+48616,48632,83815,83816,86239,86240
+48626 
+
+129973,129980,129979,129974
+
+129517,129520,129519,129518
+
+99634,99645,99644,99643,99642,99641,99640,99639,99638,99637,99636,99635
+
+40947,40948,40955,40939,40940,40956,40964,42817,46032
+
+90100,90103-90113
+
+131036,131015,132882,132889,132888,132883,132906,132913,132912,132907,129975,131066,131065,129976,132901,132900,132895,132910,134592
+
+79619,79621,79620,79622,79627,79629,79628,79630,85328,85333,85348,85332,85347,85331,85346,85361,85330,85345,85360,85365,85329,85344,85349,85364,85362,85363,86231,86232,87252,87258,87256,87254,87253,87259,87257,87255,87483,87302,88348,88349,88350,88357,88356,88355,88354,88353,88352,88351,88358,88364,88363,88362,88361,88360,88359
+
+# Some sample.  Excluded 114513 because 114840 is on the same flowcell with it, and they share
+# a barcode with identical V..J, and this barcode results in an FP between 114513 and 116096.
+# One can't solve the problem by adding 114840 to the test set because it behaves as if it's a
+# mixture of two samples.
+
+114517,114524,114521,114519,110603,106236,106237,107574,107577,107576,107575,107525,107533,107541,107542,107545,107544,107556,108167,107546,107547,107548,107557,108168,107549,107551,108162,107560,109772,110592,107520,107529,107537,110581,110583,110582,107518,107527,107534,110568,110571,110570,110569,110552,110555,110554,110553,106240,106242,106241,114510,114515,114512,110572,110575,110574,110573,110560,110563,110562,110561,108165,107555,108161,107550,107524,107532,106243,106244,110564,110567,110566,110565,107522,107531,107539,110556,110559,110558,110557,109767,109770,110588,109768,110601,110607,110602,110608,110610,110576,110577,110578,110580,110579,110597
+
+# Ovarian sample.  Excluded 123089 because it has some stuff mixed in.
+
+123085,123086,123090
+# 123089
+
+# Gastroesophageal sample.  There are two other libraries from this sample (123087, 126168)
+# however they appear to be contaminated by the ovarian sample.  The following pairwise 
+# enclone fail counts support but do not prove this.
+#
+#           123087  123088  126168  123092
+# 123085      9       0       14      0
+# 123086      9       0       15      0
+# 123089      9       0       15      0
+# 123090     11       0       12      0
+#
+# Also 123088 excluded as contaminated.
+#
+# Also 123092 excluded as contaminated.  We do not actually know that it's contaminated, however
+# enclone BCR="123085;123092" yields a very suspicious FP, as one of the barcodes is shared with
+# 123087 for the same CDR3 (barcode = GTCTCGTTCGCACTCT-1).
+
+# melanoma
+
+132884,132891,132890,132885,132896,132903,132902,132897,132908,132915,132914,132909,118053,121352,116164,116170,116169,116168,116167,116166,116165,116156,116159,116158,116157
+# 121355 should be in above but com'ed out because needs to be rerun through marsoc with 3.1
+
+# Excluded 118193 as contaminated.  What we actually see is overlap between it and
+# 99638, 99645.  Impossible to sort out with certainty.
+#
+# At one point we excluded 140365 based on putative contamination.  That was because it overlapped
+# 117417, and at the time, that was assigned to another sample.  We could now try bringing back
+# 140365, but would need to condense the data.
+
+91295-91302,91312,91314,91316,91318,91320,91322,91324,92751,92758,92763,95455,106060,106062,114844,116087-116112,117417,117703,117704,117707,118175-118192,123137,123138,123141,123142,123176,123178,123179,123182,123183,123186,123187,123190,123191,140331-140364,140366-140369
+
+# This is a superset of BI=11, probably some without matched GEX:
+
+117458-117459,129476-129486,131235,131237,131240,131256,131268,131270,131275,134532-134538
+
+86202,86216,86215,86214,86213,86212,86211,86210,86208,86207,86206,86205,86204,86203,112741,112740,113526
+# 113808 should be in above but com'ed out because needs to be rerun through marsoc with 3.1
+
+52177
+
+# This is BI=1:
+
+140696-140711
