+++ enclone_testlist/src/testdata.public.gex.human
@@ -1,0 +1,6 @@
+# some human gex datasets that are made public
+
+127798
+127801
+85679
+123217
