+++ enclone_testlist/src/testdata.public.tcr.human
@@ -1,0 +1,6 @@
+# these need to be sorted out
+
+101287
+128024
+163911
+163914
