+++ enclone_testlist/src/testdata.public.tcr.mouse
@@ -1,0 +1,1 @@
+74396
