+++ enclone_tools/Cargo.toml
@@ -32,6 +32,7 @@ edit-distance = "2"
 enclone = { git = "https://github.com/10XGenomics/enclone_ranger.git", branch = "main" }
 enclone_core = { git = "https://github.com/10XGenomics/enclone_ranger.git", branch = "main" }
 enclone_denovo = { path = "../enclone_denovo" }
+enclone_testlist = { path = "../enclone_testlist" }
 equiv = { version = "0.1", git = "https://github.com/10XGenomics/rust-toolbox.git", branch = "master" }
 evalexpr = "7"
 expr_tools = { version = "0.1", git = "https://github.com/10XGenomics/rust-toolbox.git", branch = "master" }
